Renamed deserializeLocation to deserializeLegacyLocation
